+++ crates/database/Cargo.toml
@@ -11,4 +11,3 @@ actix-web = "4.0.0-beta.19"
 chrono ={version = "0.4",features = ["serde"]}
 diesel = {git = "https://github.com/diesel-rs/diesel", features = ["chrono", "postgres","r2d2", "serde_json"]}
 serde = { version = "1", features = ["derive"] }
-validator = { version = "0.14", features = ["derive"] }
