+++ crates/database/src/dao/user.rs
@@ -1,9 +1,9 @@
-use diesel::prelude::*;
-use innocence_utils::{create_jwt, hash, Claims,Error};
 use crate::{
-    entity::{NewUser, User, UserLoginOutput},
+    entity::{NewUser, User},
     Database,
 };
+use diesel::prelude::*;
+use innocence_utils::{create_jwt, hash, Claims, Error};
 
 pub fn signup<'a>(pool: &'a Database, entity: &'a NewUser) -> Result<User, Error> {
     use crate::schema::users::dsl::*;
@@ -14,11 +14,7 @@ pub fn signup<'a>(pool: &'a Database, entity: &'a NewUser) -> Result<User, Error
         .map_err(|err| Error::DataBaseError(err.to_string()))
 }
 
-pub fn login<'a>(
-    pool: &'a Database,
-    name: &'a str,
-    psw: &'a str,
-) -> Result<UserLoginOutput, Error> {
+pub fn login<'a>(pool: &'a Database, name: &'a str, psw: &'a str) -> Result<User, Error> {
     use crate::schema::users::dsl::*;
     let hashed = hash(psw);
     let ref mut conn = pool.get()?;
@@ -26,10 +22,6 @@ pub fn login<'a>(
         .filter(username.eq(name))
         .filter(password.eq(&hashed))
         .get_result(conn)?;
-    let claims = Claims::new(user.id);
-    let user_login_output = UserLoginOutput {
-        token: create_jwt(claims)?,
-        account: user,
-    };
-    Ok(user_login_output)
+
+    Ok(user)
 }
