+++ crates/database/src/entity/bill.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Bill {
+    pub id: i32,
+    pub user: i32,
+}
